backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pillow
python-multipart
mysql-connector-python
sqlalchemy
torch
